requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-generativeai>=0.5.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
streamlit>=1.33.0
requests>=2.31.0
fpdf2>=2.7.0
